Fixed problem with respawn packet and teleporting between servers

+++ com/raphfrk/craftproxyliter/ConnectionInfo.java
@@ -56,7 +56,7 @@ public class ConnectionInfo {
 	
 	public boolean forwardConnection = false;
 	
-	public byte loginDimension = 0;
+	public int loginDimension = 0;
 	public byte loginUnknownRespawn = 0;
 	public byte loginCreative = 0;
 	public short loginHeight = 128;
